Encapsulate the CpuJob definitions inside the Job Manager
