+++ Light Vox Engine/Light Vox Engine.vcxproj
@@ -265,7 +265,6 @@ for /r "$(SolutionDir)_Binary" %%x in (*.cso) do move "%%x" "$(SolutionDir)_Bina
     <ClInclude Include="Graphics\ShaderDefinitions.h" />
     <ClInclude Include="Input\InputManager.h" />
     <ClInclude Include="JobSystem\ConcurrentQueue.h" />
-    <ClInclude Include="JobSystem\CpuJob.h" />
     <ClInclude Include="JobSystem\JobManager.h" />
     <ClInclude Include="JobSystem\JobSequence.h" />
     <ClInclude Include="Physics\Solver.h" />
